requirements.txt
fastapi
uvicorn
websockets
opencv-python
numpy
requests
rdflib
python-jose
pillow
torch
torchvision
ultralytics
